refactor: Fase 4 — storage, crash log, dynamic SD apps (F2-F4, G1-G4)
F2  os_storage API — per-app file storage on SD, path registration
F3  os_db SQLite wrapper — Kconfig-gated, espressif/idf-sqlite3 pending
F4  os_crash — logs panic/WDT/brownout reboots to /sdcard/.cyberdeck/crash.log
G1  os_manifest — cJSON parser for manifest.json (cJSON vendorizado de IDF)
G2  os_app_discover_sd — escanea /sdcard/apps/, registra apps dinámicas
G3  script stub toast — "Script runtime not available" en lugar de "Coming soon"
G4  os_script — vtable script_runtime_t para futuros intérpretes (Lua, uPy)

intent_t.app_id: uint8_t → app_id_t (uint16_t) para soportar IDs >= 256

+++ components/app_framework/app_manager.c
@@ -51,7 +51,7 @@ void app_manager_launch(app_id_t app_id, void *data, size_t data_size)
 {
     if (ui_lock(500)) {
         intent_t intent = {
-            .app_id    = (uint8_t)app_id,
+            .app_id    = app_id,
             .screen_id = 0,
             .data      = data,
             .data_size = data_size,
